freebsd: add a facter plugin to identify where LOCALBASE is
This assumes that Puppet is installed in the same place as everything
else.  Having this means that we can depend on $::localbase always
being defined on FreeBSD, so we don't have to hard-code "/usr/local"
in paths everywhere.  Should be a standard part of the FreeBSD puppet
port, which will make the resolution trivial (we can just patch the
answer in when the port is built) but may take a bit of convincing.