newsyslog: must allow for "flags" to be absent as well
newsyslog(8) allows for the flags field to be absent, so long as it is
the last thing on the line.  We cheat a bit on the regular expression
here (it would be possible to do it better but that would make it even
more difficult to understand.  We always generate a flags field.