freebsd::periodic_conf: minor robustness and convenience improvements
Add some documentation, and functionally separate the SETTING to be
updated from the resource title, in accordance with standard Puppet
behavior.  When removing a setting, allow the VALUE to be left
undefined, since it doesn't make any sense to define a value for
something that doesn't exist, but when adding a setting, make sure the
VALUE is actually a string and not some random other data type.