localbase: reference pkg package, not puppet38 package
The localbase fact is determined by checking the PREFIX of some
installed package.  We were looking for puppet38, which is clearly
wrong if you're running puppet4 (or later).  Ideally, this fact would
just be defined by the facter package itself, and would capture the
value of LOCALBASE used for port builds, but looking for where pkg
itself is installed seems like the next-best option.  (We can't look
for the facter package itself because there is also more than one of
those.)